Trait ngrams::Ngram

pub trait Ngram<'a, T: 'a + Pad + Debug + Clone>: Iterator<Item=T> where Self: Sized {
    fn ngrams(self, usize) -> Ngrams<'a, T>;
}

Iterator adaptor, allows you to call the method .ngrams(n) on your iterator, as long as the Item of the Iterator fits the trait bound

Example

use ngrams::Ngram;
let s: Vec<_> = "hello".chars().ngrams(2).collect();
assert_eq!(s, vec![
    vec!['h', 'e'],
    vec!['e', 'l'],
    vec!['l', 'l'],
    vec!['l', 'o'],
]);

Required Methods

fn ngrams(self, usize) -> Ngrams<'a, T>